+++ README.md
@@ -1,7 +1,5 @@
 # Description
 よくソースで見る関数をまとめただけ
 
-使い方は使い方.txt参照
-
 # Author
 [umeneri](https://github.com/umeneriman)
